backend/requirements.txt
# Danfoss RAG Chatbot - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# Vector Database
pinecone>=5.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Document Processing
pandas>=2.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
python-docx>=1.0.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Fuzzy String Matching
rapidfuzz>=3.0.0

# Data Processing
numpy>=1.24.0

# HTTP Client (for async operations)
httpx>=0.26.0
aiofiles>=23.0.0

# Image Processing
Pillow>=10.0.0
